Add RawSection
RawSection is pretty much the same as EmptySection except that
instead of using set_block and having it automatically generate
a palette, you have to give it a palette and a list of blocks (their
index on the palette, not a Block class).

also added EmptyRegion.add_section to make using RawSection easier.